[UPDATE|ADD] Stub
Updated the `setCurrentContent`, `getCurrentContent`, and `generateContentFile` methods related to the renaming of the `disable_file_writes` property to `disable_file_writing` in the `Crafteus\Crafteus` class.
Updated the `getData` method by adding the `key` and `default_value` arguments, and removed the return value type. This update aims to return a value based on the key sent if it is a string.
Updated the `setLastTemplating` method.
Updated the `generateContentWithTemplating` method, fixing the bug on the Templating object.
Updated the `` method by adding the `null` return type with the `?` symbol.
Added the `getTemplatingInstance` methods. `getConfig` and `getEcosystem`